+++ include/CoreGen/CoreGenCodegen/CoreGenLLVMCodegens.h
@@ -1,0 +1,21 @@
+//
+// _CoreGenLLVMCodegens_h_
+//
+// Copyright (C) 2017-2020 Tactical Computing Laboratories, LLC
+// All Rights Reserved
+// contact@tactcomplabs.com
+//
+// See LICENSE in the top level directory for licensing details
+//
+
+#ifndef _COREGENLLVMCODEGENS_H_
+#define _COREGENLLVMCODEGENS_H_
+
+// This header only includes all the derived headers for the respective
+// CoreGenLLVMCodegen code generators
+#include "CoreGen/CoreGenCodegen/CoreGenLLVMCodegen.h"
+#include "CoreGen/CoreGenCodegen/LLVMCodegen/LLVM801CG.h"
+
+#endif
+
+// EOF
