+++ src/CoreGenCodegen/CMakeLists.txt
@@ -12,14 +12,18 @@ CoreGenCodegen.cpp
 CoreGenChiselCodegen.cpp
 CoreGenLLVMCodegen.cpp
 CoreGenNodeCodegen.cpp
+CoreGenCompCodegen.cpp
 )
 
 add_subdirectory(NodeCodegen)
+add_subdirectory(LLVMCodegen)
 
 include_directories(${YAML_PATH}/include)
 include_directories(${COREGEN_INCLUDE_PATH})
 
-add_library(CoreGenCodegen ${CodegenSrcs} $<TARGET_OBJECTS:NodeCodegenObjs>)
+add_library(CoreGenCodegen ${CodegenSrcs}
+            $<TARGET_OBJECTS:NodeCodegenObjs>
+            $<TARGET_OBJECTS:LLVMCodegenObjs> )
 
 target_link_libraries(CoreGenCodegen CoreGenBackend CoreGenPluginImpl CoreGenSigMap CoreGenArchive)
