+++ src/CoreGenCodegen/LLVMCodegen/CMakeLists.txt
@@ -1,4 +1,4 @@
-# CoreGenCodegen/NodeCodegen CMakeLists.txt
+# CoreGenCodegen/LLVMCodegen CMakeLists.txt
 #
 # Copyright (C) 2017-2020 Tactical Computing Laboratories, LLC
 # All Rights Reserved
@@ -7,18 +7,13 @@
 # See LICENSE in the top level directory for licensing details
 #
 
-set(NodeCodegenSrcs
-RegClassCG.cpp
-SpadCG.cpp
-SocCG.cpp
-ISACG.cpp
-PluginCG.cpp
-CacheCG.cpp
+set(LLVMCodegenSrcs
+  LLVM801CG.cpp
 )
 
 include_directories(${YAML_PATH}/include)
 include_directories(${COREGEN_INCLUDE_PATH})
 
-add_library(NodeCodegenObjs OBJECT ${NodeCodegenSrcs})
+add_library(LLVMCodegenObjs OBJECT ${LLVMCodegenSrcs})
 
 # EOF
